docs/requirements.txt
sphinx==7.2.6
sphinx-inline-tabs
furo>=2023.9
sphinx-autodoc-typehints>=1.25,<2.0
m2r2
sphinxcontrib-mermaid
